+++ src/main/java/com/fish/fishNet/Dtos/SalidaAlimentosDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class SalidaAlimentosDto {
     private String numeroFactura;
-    private int numeroKilos;
+    private Double numeroKilos;
     private int loteId;
     private int tipoAlimentoId;
 }
